chore: address coderabbit review comments
