+++ src/components/ui/Tooltip/tests/Tooltip.test.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
-import { render, screen } from '@testing-library/react';
+import { act, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { TextEncoder, TextDecoder } from 'util';
 import Tooltip from '../Tooltip';
-;(global as any).TextEncoder = TextEncoder;
-;(global as any).TextDecoder = TextDecoder;
-const { renderToString } = require('react-dom/server');
-const { hydrateRoot } = require('react-dom/client');
-const { act } = require('react-dom/test-utils');
-
-const flush = () => new Promise(resolve => setTimeout(resolve, 0));
+import {
+    expectNoUnexpectedHydrationWarnings,
+    flush,
+    hydrateRoot,
+    renderToString
+} from '../../tests/ssrHydration';
 
 describe('Tooltip', () => {
     test('renders component', () => {
@@ -165,7 +163,7 @@ describe('Tooltip', () => {
         container.innerHTML = html;
         document.body.appendChild(container);
 
-        let root: ReturnType<typeof hydrateRoot>;
+        let root!: ReturnType<typeof hydrateRoot>;
         await act(async() => {
             root = hydrateRoot(container, (
                 <Tooltip.Root>
@@ -176,16 +174,7 @@ describe('Tooltip', () => {
             await flush();
         });
 
-        const filteredWarns = warn.mock.calls.filter(([message]) => !String(message).includes('useLayoutEffect does nothing on the server'));
-        const filteredErrors = error.mock.calls.filter(([message]) => {
-            const text = String(message);
-            return !text.includes('useLayoutEffect does nothing on the server') &&
-                !text.includes('ReactDOMTestUtils.act') &&
-                !text.includes('not wrapped in act');
-        });
-
-        expect(filteredWarns).toHaveLength(0);
-        expect(filteredErrors).toHaveLength(0);
+        expectNoUnexpectedHydrationWarnings(warn, error);
 
         await act(() => root.unmount());
         container.remove();
